myVariable: Use `long` to prevent numeric overflow

+++ myAwesomeClass.java
@@ -1,7 +1,7 @@
 public class myAwesomeClass {
 
   public static void main(String[] args) {
-    int myVariable = 15;
+    long myVariable = 15;
     for (int j = 1; j < 20; j++) {
       myVariable = myVariable * j;
       System.out.println(j + "\t" + myVariable);
